Adding hostnet and tolerations.
In this weeks episode of echo office hours I was attempting to use
flux-aio as a clusterresourceset. It worked great except that the
resulting deployment wasn't configured to use hostNetwork and didn't
tolerate the node not ready taint.

This change adds those so that flux-aio can be used in a cluster where
there is not yet a deployed cni.

Signed-off-by: Duffie Cooley <dcooley@isovalent.com>